requirements.txt
annotated-types==0.5.0
anyio==3.7.1
boto3==1.28.30
botocore==1.31.30
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
distlib==0.3.6
dnspython==2.4.1
email-validator==2.0.0.post2
fastapi==0.100.1
filelock==3.10.7
Flask==2.2.3
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
orjson==3.9.2
platformdirs==3.2.0
pydantic==2.1.1
pydantic-extra-types==2.0.0
pydantic-settings==2.0.2
pydantic_core==2.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.28.2
s3transfer==0.6.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
ujson==5.8.0
urllib3==1.26.15
uvicorn==0.23.2
uvloop==0.17.0
virtualenv==20.21.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.2.3
